+++ src/types/dictionary.ts
@@ -1,4 +1,4 @@
 import type en from "@/dictionaries/en.json";
 
 export type Dictionary = typeof en;
-export type Locale = "en" | "th";
+export type Locale = "en" | "th" | "fa" | "ar" | "ru";